+++ src/GlobalStore/timeActions.js
@@ -4,10 +4,3 @@ export function setNow(timeNow) {
         time: timeNow
     })
 }
-
-export function setSelectedTime(selectedTime) {
-    return ({
-        type: 'SET_SELECTED_TIME',
-        time: selectedTime
-    })
-}
