+++ src/LogicalFunctions/Fetch.js
