+++ src/SharedViewComponents/Slider.js
