+++ src/ViewComponents/Show.js
@@ -1,9 +1,9 @@
 import React from 'react'
 import '../App.css'
 import moment from 'moment'
-import SearchByDate from './SearchByDate'
-import SearchByEvent from './SearchByEvent'
-import SearchByTag from './SearchByTag'
+import SearchByDate from '../LogicalFunctions/SearchByDate'
+import SearchByEvent from '../LogicalFunctions/SearchByEvent'
+import SearchByTag from '../LogicalFunctions/SearchByTag'
 import store from '../GlobalStore/store'
 import { connect } from 'react-redux'
 import {setFilteredEvents} from '../GlobalStore/eventActions'
