--- src/components/MapScreen.js
@@ -1,12 +1,0 @@
-import React from 'react'
-
-const MapScreen = () => {
-
-
-    return (
-        <div>
-        </div>
-    )
-}
-
-export default MapScreen
